added alerts to a few more places, some security fixes

+++ sapphire/organizer/views.py
@@ -148,7 +148,9 @@ def addUserSlot(request, slot_id):
 
         alert = Alert(user=request.user, text="Added a volunteer openning", color=Alert.getBlue())
         alert.saveIP(request)
-
+    else:
+        alert = Alert(user=request.user, text="Only organizers can add volunteer opennings", color=Alert.getRed())
+        alert.saveIP(request)
     return redirect('/volunteer/slot/'+str(slot_id))
 
 def removeUserSlot(request, user_slot_id):
@@ -160,8 +162,12 @@ def removeUserSlot(request, user_slot_id):
         alert = Alert(user=request.user, text="Deleted a volunteer openning", color=Alert.getRed())
         alert.saveIP(request)
     else:
-        alert = Alert(user=request.user, text="A slot must have at lease 1 volunteer openning!", color=Alert.getRed())
-        alert.saveIP(request)
+        if (group.get_is_organzer(request.user) == False):
+            alert = Alert(user=request.user, text="Only organizers can delete volunteer opennings", color=Alert.getRed())
+            alert.saveIP(request)
+        else:
+            alert = Alert(user=request.user, text="A slot must have at lease 1 volunteer openning!", color=Alert.getRed())
+            alert.saveIP(request)
 
     return redirect('/volunteer/slot/'+str(user_slot.parentSlot.id))
 
@@ -201,36 +207,58 @@ def index(request):
     return redirect('/accounts/profile')
 
 def deleteEvent(request, event_id):
-    next = request.GET.get('next')
     object = Event.objects.get(id=event_id)
     group = object.parentGroup
-    name = object.name
+    if (group.get_is_organzer(request.user)):
+        name = object.name
 
-    object.delete()
+        object.delete()
 
-    feed_entry = Feed_Entry(
-        group = group,
-        user=request.user,
-        datetime=datetime.now(timezone.utc),
-        description="Deleted event \"" + name + "\"",
-        url="/volunteer/eventNeeds"
-    )
-    feed_entry.save()
+        feed_entry = Feed_Entry(
+            group = group,
+            user=request.user,
+            datetime=datetime.now(timezone.utc),
+            description="Deleted event \"" + name + "\"",
+            url="/volunteer/eventNeeds"
+        )
+        feed_entry.save()
 
-    return redirect("/volunteer/eventNeeds")
+        alert = Alert(user=request.user, text="Deleted event "+name, color=Alert.getRed())
+        alert.saveIP(request)
+
+        return redirect("/volunteer/eventNeeds")
+
+    else:
+        alert = Alert(user=request.user, text="Only organizers can delete events", color=Alert.getRed())
+        alert.saveIP(request)
+
+        return redirect("/volunteer/event/"+str(object.id))
+
 
 def deleteSlot(request, slot_id):
     slot = Slot.objects.get(id=slot_id)
-    name = slot.title
-    event = slot.parentEvent
-    slot.delete()
+    group = slot.parentEvent.parentGroup
+    if (group.get_is_organzer(request.user)):
+        name = slot.title
+        event = slot.parentEvent
+        slot.delete()
 
-    feed_entry = Feed_Entry(
-        group=event.parentGroup,
-        user=request.user,
-        datetime=datetime.now(timezone.utc),
-        description="Deleted slot \"" + name + "\" in event \"" + event.name + "\"",
-        url="/volunteer/slots"
-    )
-    feed_entry.save()
-    return redirect('/volunteer/event/'+str(event.id))
+        feed_entry = Feed_Entry(
+            group=event.parentGroup,
+            user=request.user,
+            datetime=datetime.now(timezone.utc),
+            description="Deleted slot \"" + name + "\" in event \"" + event.name + "\"",
+            url="/volunteer/slots"
+        )
+
+        feed_entry.save()
+
+        alert = Alert(user=request.user, text="Deleted slot " + name, color=Alert.getRed())
+        alert.saveIP(request)
+
+        return redirect('/volunteer/event/'+str(event.id))
+    else:
+        alert = Alert(user=request.user, text="Only organizers can delete slots", color=Alert.getRed())
+        alert.saveIP(request)
+
+        return redirect("/volunteer/slot/"+str(slot.id))
